split up opensocial tests
